fix hospital selection
