+++ frontend/src/components/Selection/DistrictSelection.tsx
@@ -1,17 +1,10 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { ChangeEvent, useCallback, useEffect } from "react";
+import { ChangeEvent, useCallback } from "react";
 import styles from "@/pages/Home/Home.module.css";
-import { getBackendUrl } from "@/utils/apiUtils";
-import {
-  useClientToken,
-  useGlobalStore,
-  useSelectionStore,
-} from "@/utils/hooks";
+import { useGlobalStore, useSelectionStore } from "@/utils/hooks";
 import { observer } from "mobx-react-lite";
-import { fetchDistricts } from "./SelectionApi";
 
 export const DistrictSelection = observer(() => {
-  const clientToken = useClientToken();
   const globalStore = useGlobalStore();
   const selectionStore = useSelectionStore();
 
@@ -22,16 +15,6 @@ export const DistrictSelection = observer(() => {
     },
     []
   );
-
-  useEffect(() => {
-    const backendUrl = getBackendUrl();
-
-    const fetchData = async () => {
-      globalStore.districts = await fetchDistricts(backendUrl, clientToken);
-    };
-
-    fetchData();
-  }, [clientToken]);
 
   return (
     <div className={styles.form_section}>
